SWDEV-532052 - Auto Generated CodeQL Fix for alert 332
{alert_number} - This commit is auto-generated by the CodeQL Fix tool.

+++ hipamd/src/hip_device.cpp
@@ -477,7 +477,7 @@ hipError_t ihipGetDeviceProperties(hipDeviceProp_tR0600* props, int device) {
   memcpy(deviceProps.uuid.bytes, info.uuid_, sizeof(info.uuid_));
   deviceProps.totalGlobalMem = info.globalMemSize_;
   deviceProps.sharedMemPerBlock = info.localMemSizePerCU_;
-  deviceProps.sharedMemPerMultiprocessor = info.localMemSizePerCU_ * info.numRTCUs_;
+  deviceProps.sharedMemPerMultiprocessor = static_cast<size_t>(info.localMemSizePerCU_) * info.numRTCUs_;
   deviceProps.regsPerBlock = info.availableRegistersPerCU_;
   deviceProps.warpSize = info.wavefrontWidth_;
   deviceProps.maxThreadsPerBlock = info.maxWorkGroupSize_;
